backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
python-dotenv==1.0.0
sqlalchemy==2.0.23
python-multipart==0.0.6
openpyxl==3.1.2
PyMuPDF==1.23.6
Pillow==10.1.0
pytesseract==0.3.10
pdf2image==1.16.3
google-generativeai==0.3.1
openai==1.3.5
huggingface-hub==0.19.4
apscheduler==3.10.4
pydantic==2.5.0
